feat(kikan): V3 — lift platform DB init + PRAGMA setup from mokumo_db [#507]
Stage 3 Session S1.1: move pool creation, PRAGMA configuration, backup,
and restore-candidate validation into the `kikan::db` + `kikan::backup`
module trees. This resolves the Stage 0.2 "caller provides" punt —
`initialize_database` is now pool-only, with migrations kept as the
vertical's responsibility per the Migration Ownership golden rule.

- `kikan::db::{pragmas, application_id, init}` — lifted verbatim from
  `crates/db/src/lib.rs` with two surface changes:
  - `MOKUMO_APPLICATION_ID` renamed to `KIKAN_APPLICATION_ID`. Stored
    byte value frozen at `0x4D4B4D4F` ("MKMO"); a constant-byte-value
    assertion guards pre-Stage-3 profile DBs against drift.
  - `initialize_database` is pool-only (no Migrator::up). `check_schema_
    compatibility<M: MigratorTrait>` is generic so kikan has no
    dependency on a specific vertical's migrator.
- `kikan::backup::{snapshot, restore}` — lifted from `crates/db/src/
  {backup.rs, restore.rs}`. `validate_candidate<M>` is generic;
  `copy_to_production` takes the production filename as a parameter so
  the primitive stays vertical-agnostic.
- `crates/db/src/backup.rs` is deleted; `crates/db/src/restore.rs`
  becomes a 30-line wrapper binding the generic primitive to
  `migration::Migrator` and `"mokumo.db"`.
- `mokumo_db::initialize_database` stays as a thin vertical wrapper
  that runs the mokumo migrator + downgrade-detection on top of the
  kikan pool opener. Keeps `crates/db/tests/*` compiling; both the
  wrapper and the crate dissolve in S3.1b.
- `services/api` call sites for pure-platform symbols (pool, backup
  primitives, application-id and auto-vacuum guards) migrate to
  `kikan::db::*` / `kikan::backup::*` directly.

I1–I5 scripts pass; `cargo test -p kikan -p mokumo-db` green.
